refactor: simplify attack validation logic, remove redundant code, and add comprehensive unit tests

+++ src/Core/NeoServer.Domain/Common/Location/TileFlag.cs
@@ -1,6 +1,6 @@
 ﻿namespace NeoServer.Domain.Common.Location;
 
-public enum TileFlag : byte
+public enum TileFlag : uint
 {
     None = 0
 }
